Include an SPI device FPGA logic example
This added v/comm.v example demonstrates how the RP2350B can use
SPI over the ICE_spi pins (pico2ice.SPIxF(write, read)) to
write/read bytes to/from logic on the FPGA. Extended the
v/hello.pcf file to include reference to the ICE_spi pins.

Increased the clock speed of the FPGA logic by a factor of 3 to
75 MHz by modifying pio/clock.pio and rebuilt the pio.go code
to use that. For the comm.v code to work, the FPGA logic clock
has to be 8+ times faster than the SPI clock.

Signed-off-by: Tinkerer <tinkerer@zappem.net>